moved functions to functions.ts
